requirements.txt
Django~=4.1
djangorestframework~=3.13.1
uuid~=1.30
gunicorn==20.1.0